rename interface to BundleReader, move to mindpak pkg
